chg: 🔥 Fix and refactor cypress tests
