--- cypress/declarations.d.ts
@@ -1,6 +1,0 @@
-declare module '*.png'
-declare module '*.jpeg'
-declare module '*.pdf'
-declare module '*.jpg'
-declare module '*.jpeg'
-declare module '*.svg'
