+++ cypress/support/index.d.ts
@@ -1,0 +1,9 @@
+import { mount } from 'cypress/react18'
+
+declare global {
+  namespace Cypress {
+    interface Chainable {
+      mount: typeof mount
+    }
+  }
+}
