+++ package.json
@@ -40,7 +40,7 @@
     "typescript-transform-paths": "^3.3.1",
     "vite": "^4.3.9",
     "vite-plugin-dts": "^3.0.3",
-    "vite-tsconfig-paths": "^3.3.13"
+    "vite-tsconfig-paths": "^4.2.0"
   },
   "peerDependencies": {
     "react": "^17.0.2"
