+++ tsconfig.json
@@ -10,16 +10,10 @@
     "esModuleInterop": true,
     "baseUrl": ".",
     "paths": {
-      "@/*": ["src/*"]
-    },
-    "plugins": [
-      // Transform paths in output .js files
-      { "transform": "typescript-transform-paths" },
-
-      // Transform paths in output .d.ts files (Include this line if you output declarations files)
-      { "transform": "typescript-transform-paths", "afterDeclarations": true }
-    ]
+      "@/*": ["src/*"],
+      "cypress/*": ["./cypress/*"]
+    }
   },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist", "docs", "src/**/*.spec.tsx", "cypress"]
+  "include": ["src", "cypress"],
+  "exclude": ["node_modules", "dist", "docs"]
 }
